+++ src/main/java/nextstep/subway/auth/ui/AuthController.java
@@ -1,12 +1,9 @@
 package nextstep.subway.auth.ui;
 
 import nextstep.subway.auth.application.AuthService;
-import nextstep.subway.auth.application.AuthorizationException;
 import nextstep.subway.auth.dto.TokenRequest;
 import nextstep.subway.auth.dto.TokenResponse;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -25,8 +22,5 @@ public class AuthController {
         return ResponseEntity.ok().body(token);
     }
 
-    @ExceptionHandler(AuthorizationException.class)
-    public ResponseEntity handleAuthorizationException(Exception e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
+
 }
