+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -17,12 +17,11 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-@Transactional(readOnly = true)
+@Transactional
 public class FavoriteService {
 
     private final String MESSAGE_FAVORITE_ENTITY_NOT_FOUND = "즐겨찾기가가 존재하지 않습니다";
-
-
+    
     private FavoriteRepository favoriteRepository;
     private MemberService memberService;
     private StationService stationService;
@@ -59,6 +58,7 @@ public class FavoriteService {
         pathService.findShortestPath(sourceStationId, targetStationId);
     }
 
+    @Transactional(readOnly = true)
     public List<FavoriteResponse> findFavorites(Long memberId) {
         Member member = findMember(memberId);
         List<Favorite> favorites = favoriteRepository.findByMemberId(member.getId());
